refactor: vod 도메인 클래스 이관
- domain의 stream 패키지로 이동
