+++ src/main/java/com/ssginc/showpingrefactoring/domain/stream/infrastructure/subtitle/SubtitleGenerator.java
@@ -1,4 +1,4 @@
-package com.ssginc.showpingrefactoring.vod.infrastructure.subtitle;
+package com.ssginc.showpingrefactoring.domain.stream.infrastructure.subtitle;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
