+++ src/main/java/com/ssginc/showpingrefactoring/domain/stream/repository/VodRepository.java
@@ -1,6 +1,7 @@
-package com.ssginc.showpingrefactoring.vod.repository;
+package com.ssginc.showpingrefactoring.domain.stream.repository;
 
-import com.ssginc.showpingrefactoring.vod.dto.response.VodResponseDto;
+import com.ssginc.showpingrefactoring.domain.stream.dto.response.VodResponseDto;
+import com.ssginc.showpingrefactoring.domain.stream.entity.Stream;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
